Move time slots to screens folder
Refs #435.
