+++ app/assets/styles/app.less
@@ -41,7 +41,6 @@
 @import './reservations-filters';
 @import './reservations-list';
 @import './resources-table';
-@import './time-slots';
 @import './typeahead-list';
 @import './wrapped-text';
