+++ app/screens/resource/reservation-calendar/time-slots/TimeSlots.spec.js
@@ -1,13 +1,12 @@
 import { expect } from 'chai';
 import React from 'react';
-import sd from 'skin-deep';
-import simple from 'simple-mock';
-
 import Immutable from 'seamless-immutable';
+import simple from 'simple-mock';
+import sd from 'skin-deep';
 
-import TimeSlots from 'components/reservation/TimeSlots';
 import Resource from 'fixtures/Resource';
 import TimeSlot from 'fixtures/TimeSlot';
+import TimeSlots from './TimeSlots';
 
 describe('Component: reservation/TimeSlots', () => {
   describe('with timeslots', () => {
