+++ app/screens/resource/reservation-calendar/time-slots/index.js
@@ -1,0 +1,3 @@
+import TimeSlots from './TimeSlots';
+
+export default TimeSlots;
