+++ app/screens/resource/reservation-calendar/time-slots/time-slots.less
