+++ app/screens/resource/resource-page.less
@@ -1,1 +1,2 @@
+@import './reservation-calendar/time-slots/time-slots';
 @import './resource-info/resource-info';
